requirements.txt
numpy
scipy
numpy-stl
xarray
py3dep